RoseEngine

Documentation for RoseEngine.

RoseEngine.Rotor

RoseEngine.full_cycle_count

RoseEngine.triangle_wave

RoseEngine.Rotor — Type

Rotor(radius, amplitude, waves_per_rotation)

Rotor is a function that for a given angle, returns a radius from the center of the Rotor's spindle. In this package, angles are in units of fractions of a full circle, arther than in degrees or radians.

radius is the radius at zero amplitude.

amplitude is the extent above and below radius of the range of the "cutter". The range of the Rotor funnction will be from radius - amplitude to radius + amplitude.

wave_function is a function describing one cycle of the wave. Its domain goes from 0 to 1 and its range goes from -1 to 1. See [triangle_wave(@ref) as an example.

waves_per_rotation is the number of full oscillations of wave_function per one turn of the Rotor.

RoseEngine.full_cycle_count — Method

full_cycle_count(::Rotor)::Int

Returns the number of cycles that the Rotor should rotate such that its wave_function returns to its starting point (same angle and radius). This is an integer number of full rotations of the rootor.

RoseEngine.triangle_wave — Method

triangle_wave(p)

triangle_wave implements a symetric triangle wave suitable for use as the wave_function of Rotor.